requirements.txt
streamlit
pymongo
openai
python-dotenv
yfinance
pandas
numpy
torch
transformers
matplotlib